refact, rename to pageFocus & Wrapper, rm Today/Belong into /remain.
