+++ app/Within/pageFocus/Main/Broads/Broads.jsx
@@ -5,7 +5,7 @@ import {
 } from 'react-router-dom';
 import {connect} from "react-redux";
 import classnames from 'classnames';
-import styles from "../MainBanner/styles.module.css"; //Notice, we use shared css file here for easier control
+import styles from "../styles.module.css"; //Notice, we use shared css file here for easier control
 import {
   axios_Units,
 } from '../MainBanner/utils.js';
