+++ app/Within/pageFocus/Main/MainBanner/utils.js
