+++ app/Within/pageFocus/Main/MainTitle/MainTitle.jsx
@@ -9,7 +9,6 @@ import {
 import {connect} from "react-redux";
 import classnames from 'classnames';
 import styles from "./styles.module.css";
-import BelongForm from '../../component/BelongForm/BelongForm.jsx';
 import DateConverter from '../../../Component/DateConverter.jsx';
 import CreateShare from '../../../Component/CreateShare.jsx';
 import SvgCreate from '../../../Component/Svg/SvgCreate.jsx';
@@ -102,10 +101,7 @@ class MainTitle extends React.Component {
             <div
               style={{height: "100%", position: 'absolute', right: '-6%', top: '0',borderRight: 'solid 1px #6e6e6e'}}/>
           </div>
-          <div
-            className={classnames(styles.boxBelong)}>
-            <BelongForm/>
-          </div>
+
         </div>
 
       </div>
